requirements.txt
PySide6~=6.10.0
requests~=2.32.5
python-dotenv~=1.1.1
